Updated to parse GENE_TRANSCRIPT_XREF format in INFO section
Also changed file open funciton to use gzip if the file ends in .gz

Changed the output file name to replace only the last instance of .vcf to .json

Included turning entries with an & in them into an array in the json output.

Had some issues with the length of the array with the format_tag and format not being there in
my file so I included some logic to check the length of the split tag first.

Lastly, pretty printed the output JSON using indents.